requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==3.4.8
distlib==0.3.4
filelock==3.4.0
Flask==1.1.2
Flask-Cors==3.0.10
flask-pyjwt==0.1.9
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pipenv==2021.11.23
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.17.1
pytz==2021.3
six==1.16.0
toml==0.10.2
virtualenv==20.10.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
